move verification primitives to main module
